Redistribute advanced settings
