--- src/components/settings/AdvancedSettingsPane.tsx
@@ -1,582 +1,0 @@
-import { InformationCircleIcon } from "@hugeicons/core-free-icons";
-import { HugeiconsIcon } from "@hugeicons/react";
-import { useCallback, useEffect, useState } from "react";
-import { useSpace } from "../../contexts";
-import { useFileTreeSortMode } from "../../hooks/useFileTreeSortMode";
-import { extractErrorMessage } from "../../lib/errorUtils";
-import {
-	type AiAssistantMode,
-	type EditorWidthMode,
-	FILE_TREE_SORT_OPTIONS,
-	isFileTreeSortMode,
-	loadSettings,
-	setAiAssistantMode,
-	setClassicAllNotesByDefault,
-	setDatabaseShowColumnColor,
-	setEditorBeautifulTags,
-	setEditorColorfulHeadings,
-	setEditorEnablePeopleMentionsAsTags,
-	setEditorShowCollapsibleHeadings,
-	setEditorShowFrontmatterInEditor,
-	setEditorSpellCheck,
-	setEditorVimKeybindings,
-	setEditorWidthMode,
-	setFolioMode,
-	setShowFileTreeFolderCounts,
-	setShowNonMarkdownFiles,
-	setShowToc,
-} from "../../lib/settings";
-import { invoke } from "../../lib/tauri";
-import { useTauriEvent } from "../../lib/tauriEvents";
-import { Popover, PopoverContent, PopoverTrigger } from "../ui/shadcn/popover";
-import {
-	SettingsRow,
-	SettingsSection,
-	SettingsToggle,
-} from "./SettingsScaffold";
-import { SettingsSelect } from "./SettingsSelect";
-import { useOptimisticSettingsToggle } from "./useOptimisticSettingsToggle";
-
-const VIM_KEYBINDING_HELP = [
-	{ key: "Esc", action: "Enter Vim command mode." },
-	{ key: "i", action: "Type at the cursor." },
-	{ key: "a", action: "Type after the cursor." },
-	{ key: "I", action: "Go to the start of the line and type." },
-	{ key: "A", action: "Go to the end of the line and type." },
-	{ key: "o", action: "Open a new line below and type." },
-	{ key: "O", action: "Open a new line above and type." },
-	{ key: "h / j / k / l", action: "Move left, down, up, and right." },
-	{ key: "w", action: "Jump to the next word." },
-	{ key: "b", action: "Jump back to the previous word." },
-	{ key: "e", action: "Jump to the end of the word." },
-	{ key: "0", action: "Jump to the start of the line." },
-	{ key: "$", action: "Jump to the end of the line." },
-	{ key: "gg", action: "Jump to the start of the note." },
-	{ key: "G", action: "Jump to the end of the note." },
-	{ key: "x", action: "Delete the character under or near the cursor." },
-	{ key: "dd", action: "Delete the current line's contents." },
-	{ key: "u", action: "Undo." },
-	{ key: "Control-r", action: "Redo." },
-] as const;
-
-const EDITOR_WIDTH_OPTIONS = [
-	{ label: "Compact", value: "compact" },
-	{ label: "Comfortable", value: "comfortable" },
-	{ label: "Wide", value: "wide" },
-] as const satisfies readonly { label: string; value: EditorWidthMode }[];
-
-function VimKeybindingsHelp() {
-	return (
-		<Popover>
-			<PopoverTrigger asChild>
-				<button
-					type="button"
-					className="vimKeybindingsInfoButton"
-					aria-label="Vim keybindings help"
-				>
-					<HugeiconsIcon
-						icon={InformationCircleIcon}
-						size="var(--icon-md)"
-						strokeWidth={0.9}
-					/>
-				</button>
-			</PopoverTrigger>
-			<PopoverContent
-				align="start"
-				side="right"
-				sideOffset={8}
-				className="vimKeybindingsPopover"
-			>
-				<div className="vimKeybindingsPopoverTitle">Vim keybindings</div>
-				<div className="vimKeybindingsModes">
-					<div>
-						<strong>insert</strong> means normal typing mode. You type and text
-						appears, like the editor already does.
-					</div>
-					<div>
-						<strong>normal</strong> means command mode. Your keys move around or
-						edit text instead of typing letters.
-					</div>
-				</div>
-				<div className="vimKeybindingsList">
-					{VIM_KEYBINDING_HELP.map((item) => (
-						<div className="vimKeybindingsItem" key={item.key}>
-							<kbd>{item.key}</kbd>
-							<span>{item.action}</span>
-						</div>
-					))}
-				</div>
-			</PopoverContent>
-		</Popover>
-	);
-}
-
-export function AdvancedSettingsPane() {
-	const [showCollapsibleHeadings, setShowCollapsibleHeadings] = useState(false);
-	const [showFrontmatterInEditor, setShowFrontmatterInEditor] = useState(false);
-	const [colorfulHeadings, setColorfulHeadings] = useState(false);
-	const [beautifulTags, setBeautifulTags] = useState(false);
-	const [editorWidthMode, setEditorWidthModeState] =
-		useState<EditorWidthMode>("compact");
-	const [enablePeopleMentionsAsTags, setEnablePeopleMentionsAsTags] =
-		useState(false);
-	const [vimKeybindings, setVimKeybindings] = useState(false);
-	const [spellCheck, setSpellCheck] = useState(true);
-	const [showToc, setShowTocState] = useState(true);
-	const [aiAssistantMode, setAiAssistantModeState] =
-		useState<AiAssistantMode>("create");
-	const [folioMode, setFolioModeState] = useState(false);
-	const [classicAllNotesByDefault, setClassicAllNotesByDefaultState] =
-		useState(false);
-	const [showFileTreeFolderCounts, setShowFileTreeFolderCountsState] =
-		useState(false);
-	const [showNonMarkdownFiles, setShowNonMarkdownFilesState] = useState(true);
-	const [showDatabaseColumnColor, setShowDatabaseColumnColor] = useState(true);
-	const [error, setError] = useState("");
-	const [
-		isSavingEnablePeopleMentionsAsTags,
-		setIsSavingEnablePeopleMentionsAsTags,
-	] = useState(false);
-	const [isSavingEditorWidthMode, setIsSavingEditorWidthMode] = useState(false);
-	const [isSavingAiAssistantMode, setIsSavingAiAssistantMode] = useState(false);
-	const { spacePath, startIndexRebuild } = useSpace();
-	const fileTreeSort = useFileTreeSortMode({
-		onError: setError,
-	});
-	const showTocToggle = useOptimisticSettingsToggle(
-		showToc,
-		setShowTocState,
-		setShowToc,
-		setError,
-	);
-	const showFrontmatterToggle = useOptimisticSettingsToggle(
-		showFrontmatterInEditor,
-		setShowFrontmatterInEditor,
-		setEditorShowFrontmatterInEditor,
-		setError,
-	);
-	const colorfulHeadingsToggle = useOptimisticSettingsToggle(
-		colorfulHeadings,
-		setColorfulHeadings,
-		setEditorColorfulHeadings,
-		setError,
-	);
-	const beautifulTagsToggle = useOptimisticSettingsToggle(
-		beautifulTags,
-		setBeautifulTags,
-		setEditorBeautifulTags,
-		setError,
-	);
-	const showCollapsibleHeadingsToggle = useOptimisticSettingsToggle(
-		showCollapsibleHeadings,
-		setShowCollapsibleHeadings,
-		setEditorShowCollapsibleHeadings,
-		setError,
-	);
-	const spellCheckToggle = useOptimisticSettingsToggle(
-		spellCheck,
-		setSpellCheck,
-		setEditorSpellCheck,
-		setError,
-	);
-	const vimKeybindingsToggle = useOptimisticSettingsToggle(
-		vimKeybindings,
-		setVimKeybindings,
-		setEditorVimKeybindings,
-		setError,
-	);
-	const folioModeToggle = useOptimisticSettingsToggle(
-		folioMode,
-		setFolioModeState,
-		setFolioMode,
-		setError,
-	);
-	const classicAllNotesToggle = useOptimisticSettingsToggle(
-		classicAllNotesByDefault,
-		setClassicAllNotesByDefaultState,
-		setClassicAllNotesByDefault,
-		setError,
-	);
-	const showFileTreeFolderCountsToggle = useOptimisticSettingsToggle(
-		showFileTreeFolderCounts,
-		setShowFileTreeFolderCountsState,
-		setShowFileTreeFolderCounts,
-		setError,
-	);
-	const showNonMarkdownFilesToggle = useOptimisticSettingsToggle(
-		showNonMarkdownFiles,
-		setShowNonMarkdownFilesState,
-		setShowNonMarkdownFiles,
-		setError,
-	);
-	const showDatabaseColumnColorToggle = useOptimisticSettingsToggle(
-		showDatabaseColumnColor,
-		setShowDatabaseColumnColor,
-		setDatabaseShowColumnColor,
-		setError,
-	);
-
-	const refresh = useCallback(async () => {
-		setError("");
-		try {
-			const settings = await loadSettings();
-			setShowCollapsibleHeadings(settings.editor.showCollapsibleHeadings);
-			setShowFrontmatterInEditor(settings.editor.showFrontmatterInEditor);
-			setColorfulHeadings(settings.editor.colorfulHeadings);
-			setBeautifulTags(settings.editor.beautifulTags);
-			setEditorWidthModeState(settings.editor.editorWidthMode);
-			setEnablePeopleMentionsAsTags(settings.editor.enablePeopleMentionsAsTags);
-			setVimKeybindings(settings.editor.vimKeybindings === true);
-			setSpellCheck(settings.editor.spellCheck !== false);
-			setShowTocState(settings.ui.showToc);
-			setAiAssistantModeState(settings.ui.aiAssistantMode);
-			setFolioModeState(settings.ui.folioMode);
-			setClassicAllNotesByDefaultState(settings.ui.classicAllNotesByDefault);
-			setShowFileTreeFolderCountsState(settings.ui.showFileTreeFolderCounts);
-			setShowNonMarkdownFilesState(settings.ui.showNonMarkdownFiles);
-			setShowDatabaseColumnColor(settings.database.showColumnColor);
-		} catch (cause) {
-			setError(extractErrorMessage(cause));
-		}
-	}, []);
-
-	useEffect(() => {
-		void refresh();
-	}, [refresh]);
-
-	useTauriEvent("settings:updated", (payload) => {
-		if (typeof payload.editor?.showCollapsibleHeadings === "boolean") {
-			setShowCollapsibleHeadings(payload.editor.showCollapsibleHeadings);
-		}
-		if (typeof payload.editor?.showFrontmatterInEditor === "boolean") {
-			setShowFrontmatterInEditor(payload.editor.showFrontmatterInEditor);
-		}
-		if (typeof payload.editor?.colorfulHeadings === "boolean") {
-			setColorfulHeadings(payload.editor.colorfulHeadings);
-		}
-		if (typeof payload.editor?.beautifulTags === "boolean") {
-			setBeautifulTags(payload.editor.beautifulTags);
-		}
-		if (
-			payload.editor?.editorWidthMode === "compact" ||
-			payload.editor?.editorWidthMode === "comfortable" ||
-			payload.editor?.editorWidthMode === "wide"
-		) {
-			setEditorWidthModeState(payload.editor.editorWidthMode);
-		}
-		if (typeof payload.editor?.enablePeopleMentionsAsTags === "boolean") {
-			setEnablePeopleMentionsAsTags(payload.editor.enablePeopleMentionsAsTags);
-		}
-		if (typeof payload.editor?.vimKeybindings === "boolean") {
-			setVimKeybindings(payload.editor.vimKeybindings);
-		}
-		if (typeof payload.editor?.spellCheck === "boolean") {
-			setSpellCheck(payload.editor.spellCheck);
-		}
-		if (typeof payload.ui?.showToc === "boolean") {
-			setShowTocState(payload.ui.showToc);
-		}
-		if (
-			payload.ui?.aiAssistantMode === "chat" ||
-			payload.ui?.aiAssistantMode === "create"
-		) {
-			setAiAssistantModeState(payload.ui.aiAssistantMode);
-		}
-		if (typeof payload.ui?.folioMode === "boolean") {
-			setFolioModeState(payload.ui.folioMode);
-		}
-		if (typeof payload.ui?.classicAllNotesByDefault === "boolean") {
-			setClassicAllNotesByDefaultState(payload.ui.classicAllNotesByDefault);
-		}
-		if (typeof payload.ui?.showFileTreeFolderCounts === "boolean") {
-			setShowFileTreeFolderCountsState(payload.ui.showFileTreeFolderCounts);
-		}
-		if (typeof payload.ui?.showNonMarkdownFiles === "boolean") {
-			setShowNonMarkdownFilesState(payload.ui.showNonMarkdownFiles);
-		}
-		if (typeof payload.database?.showColumnColor === "boolean") {
-			setShowDatabaseColumnColor(payload.database.showColumnColor);
-		}
-	});
-
-	return (
-		<div className="settingsPane">
-			{error ? <div className="settingsError">{error}</div> : null}
-
-			<div className="settingsGrid">
-				<SettingsSection
-					title="Editor"
-					description="Controls for editing behavior and note structure inside Glyph."
-				>
-					<SettingsRow
-						label="Table of contents"
-						description="Show a floating table of contents for each note."
-					>
-						<SettingsToggle
-							checked={showToc}
-							disabled={showTocToggle.isSaving}
-							ariaLabel="Table of contents"
-							onCheckedChange={showTocToggle.onCheckedChange}
-						/>
-					</SettingsRow>
-					<SettingsRow
-						label="People mentions as tags"
-						description="When enabled, standalone @mentions are indexed and shown as people in the sidebar and search."
-					>
-						<SettingsToggle
-							checked={enablePeopleMentionsAsTags}
-							disabled={isSavingEnablePeopleMentionsAsTags}
-							ariaLabel="People mentions as tags"
-							onCheckedChange={(checked) => {
-								const previous = enablePeopleMentionsAsTags;
-								setError("");
-								setIsSavingEnablePeopleMentionsAsTags(true);
-								void (async () => {
-									await invoke("index_set_people_mentions_as_tags_enabled", {
-										enabled: checked,
-									});
-									if (spacePath) {
-										await startIndexRebuild();
-									}
-									await setEditorEnablePeopleMentionsAsTags(checked);
-									setEnablePeopleMentionsAsTags(checked);
-								})()
-									.catch((cause) => {
-										setEnablePeopleMentionsAsTags(previous);
-										void invoke("index_set_people_mentions_as_tags_enabled", {
-											enabled: previous,
-										}).catch(() => undefined);
-										setError(extractErrorMessage(cause));
-									})
-									.finally(() => {
-										setIsSavingEnablePeopleMentionsAsTags(false);
-									});
-							}}
-						/>
-					</SettingsRow>
-					<SettingsRow
-						label="Show frontmatter in editor"
-						description="Display YAML frontmatter at the top of notes while editing. Turning this off keeps frontmatter available to indexing and databases."
-					>
-						<SettingsToggle
-							checked={showFrontmatterInEditor}
-							disabled={showFrontmatterToggle.isSaving}
-							ariaLabel="Show frontmatter in editor"
-							onCheckedChange={showFrontmatterToggle.onCheckedChange}
-						/>
-					</SettingsRow>
-					<SettingsRow
-						label="Colorful headings"
-						description="Use distinct built-in colors for H1-H6 while editing notes."
-					>
-						<SettingsToggle
-							checked={colorfulHeadings}
-							disabled={colorfulHeadingsToggle.isSaving}
-							ariaLabel="Colorful headings"
-							onCheckedChange={colorfulHeadingsToggle.onCheckedChange}
-						/>
-					</SettingsRow>
-					<SettingsRow
-						label="Beautiful Tags"
-						description="Enable the experimental Beautiful Tags presentation for tags."
-					>
-						<SettingsToggle
-							checked={beautifulTags}
-							disabled={beautifulTagsToggle.isSaving}
-							ariaLabel="Beautiful Tags"
-							onCheckedChange={beautifulTagsToggle.onCheckedChange}
-						/>
-					</SettingsRow>
-					<SettingsRow
-						label="Editor width"
-						description="Compact keeps lines shorter, Comfortable gives a little bit more room, and Wide uses the full editor width."
-						interactive={false}
-					>
-						<SettingsSelect
-							aria-label="Editor width"
-							value={editorWidthMode}
-							disabled={isSavingEditorWidthMode}
-							onChange={(event) => {
-								const nextMode = event.currentTarget.value as EditorWidthMode;
-								const previous = editorWidthMode;
-								setError("");
-								setEditorWidthModeState(nextMode);
-								setIsSavingEditorWidthMode(true);
-								void setEditorWidthMode(nextMode)
-									.catch((cause) => {
-										setEditorWidthModeState(previous);
-										setError(extractErrorMessage(cause));
-									})
-									.finally(() => {
-										setIsSavingEditorWidthMode(false);
-									});
-							}}
-						>
-							{EDITOR_WIDTH_OPTIONS.map((option) => (
-								<option key={option.value} value={option.value}>
-									{option.label}
-								</option>
-							))}
-						</SettingsSelect>
-					</SettingsRow>
-					<SettingsRow
-						label="Collapsible headings"
-						description="Show collapse toggles on note headings in editor and preview."
-					>
-						<SettingsToggle
-							checked={showCollapsibleHeadings}
-							disabled={showCollapsibleHeadingsToggle.isSaving}
-							ariaLabel="Collapsible headings"
-							onCheckedChange={showCollapsibleHeadingsToggle.onCheckedChange}
-						/>
-					</SettingsRow>
-					<SettingsRow
-						label="Spell check"
-						description="Underline typos as you type. Right-click a word to see spelling suggestions."
-					>
-						<SettingsToggle
-							checked={spellCheck}
-							disabled={spellCheckToggle.isSaving}
-							ariaLabel="Spell check"
-							onCheckedChange={spellCheckToggle.onCheckedChange}
-						/>
-					</SettingsRow>
-					<SettingsRow
-						label={
-							<span className="settingsLabelWithHelp">
-								Vim Mode
-								<VimKeybindingsHelp />
-							</span>
-						}
-						description="Do NOT Turn this ON if you don't know what it means."
-					>
-						<SettingsToggle
-							checked={vimKeybindings}
-							disabled={vimKeybindingsToggle.isSaving}
-							ariaLabel="Vim Mode"
-							onCheckedChange={vimKeybindingsToggle.onCheckedChange}
-						/>
-					</SettingsRow>
-				</SettingsSection>
-				<SettingsSection
-					title="AI"
-					description="Controls for how the AI assistant behaves by default."
-				>
-					<SettingsRow
-						label="AI chat has access to tools"
-						description="When on, AI can use tools to create and take actions. When off, it stays in chat-only mode."
-					>
-						<SettingsToggle
-							checked={aiAssistantMode === "create"}
-							disabled={isSavingAiAssistantMode}
-							ariaLabel="AI chat has access to tools"
-							onCheckedChange={(checked) => {
-								const previous = aiAssistantMode;
-								const nextMode: AiAssistantMode = checked ? "create" : "chat";
-								setError("");
-								setAiAssistantModeState(nextMode);
-								setIsSavingAiAssistantMode(true);
-								void setAiAssistantMode(nextMode)
-									.catch((cause) => {
-										setAiAssistantModeState(previous);
-										setError(extractErrorMessage(cause));
-									})
-									.finally(() => {
-										setIsSavingAiAssistantMode(false);
-									});
-							}}
-						/>
-					</SettingsRow>
-				</SettingsSection>
-				<SettingsSection
-					title="App"
-					description="Global app-level controls for the sidebar and workspace UI."
-				>
-					<SettingsRow
-						label="Folio Mode"
-						description="Show navigation, notes, and editor in a three-column workspace."
-					>
-						<SettingsToggle
-							checked={folioMode}
-							disabled={folioModeToggle.isSaving}
-							ariaLabel="Folio Mode"
-							onCheckedChange={folioModeToggle.onCheckedChange}
-						/>
-					</SettingsRow>
-					<SettingsRow
-						label="Classic All Notes grid"
-						description="Open All Notes as the simple grid instead of the activity timeline."
-					>
-						<SettingsToggle
-							checked={classicAllNotesByDefault}
-							disabled={classicAllNotesToggle.isSaving}
-							ariaLabel="Classic All Notes grid"
-							onCheckedChange={classicAllNotesToggle.onCheckedChange}
-						/>
-					</SettingsRow>
-					<SettingsRow
-						label="Show folder file counts"
-						description="Show a recursive file total at the end of each folder row in the file tree."
-					>
-						<SettingsToggle
-							checked={showFileTreeFolderCounts}
-							disabled={showFileTreeFolderCountsToggle.isSaving}
-							ariaLabel="Show folder file counts"
-							onCheckedChange={showFileTreeFolderCountsToggle.onCheckedChange}
-						/>
-					</SettingsRow>
-					<SettingsRow
-						label="Show non-Markdown files"
-						description="Show PDFs, images, and other attachments in the file tree and Folio list. Turning this off hides them from those views only."
-					>
-						<SettingsToggle
-							checked={showNonMarkdownFiles}
-							disabled={showNonMarkdownFilesToggle.isSaving}
-							ariaLabel="Show non-Markdown files"
-							onCheckedChange={showNonMarkdownFilesToggle.onCheckedChange}
-						/>
-					</SettingsRow>
-					<SettingsRow
-						label="File tree sort"
-						description="Choose how folders and files are ordered in the sidebar tree."
-						interactive={false}
-					>
-						<SettingsSelect
-							aria-label="File tree sort"
-							value={fileTreeSort.sortMode}
-							disabled={fileTreeSort.isSaving}
-							onChange={(event) => {
-								const nextSortMode = event.currentTarget.value;
-								if (!isFileTreeSortMode(nextSortMode)) return;
-								setError("");
-								void fileTreeSort.setSortMode(nextSortMode);
-							}}
-						>
-							{FILE_TREE_SORT_OPTIONS.map((option) => (
-								<option key={option.value} value={option.value}>
-									{option.label}
-								</option>
-							))}
-						</SettingsSelect>
-					</SettingsRow>
-				</SettingsSection>
-				<SettingsSection
-					title="Database"
-					description="Global controls for how databases are presented across Glyph."
-				>
-					<SettingsRow
-						label="Show database column color"
-						description="Keep the lane pill and tag colors while toggling the full column tint."
-					>
-						<SettingsToggle
-							checked={showDatabaseColumnColor}
-							disabled={showDatabaseColumnColorToggle.isSaving}
-							ariaLabel="Show database column color"
-							onCheckedChange={showDatabaseColumnColorToggle.onCheckedChange}
-						/>
-					</SettingsRow>
-				</SettingsSection>
-			</div>
-		</div>
-	);
-}
