+++ src/components/settings/settingsSearch.ts
@@ -454,99 +454,113 @@ const SETTINGS_SEARCH_ITEMS = [
 		keywords: ["sync", "files"],
 	},
 	{
-		id: "advanced-table-of-contents",
-		tab: "advanced",
+		id: "general-editor-table-of-contents",
+		tab: "general",
 		section: "Editor",
 		title: "Table of contents",
 		keywords: ["toc", "outline"],
 	},
 	{
-		id: "advanced-people-tags",
-		tab: "advanced",
-		section: "Editor",
+		id: "space-search-index-people-tags",
+		tab: "space",
+		section: "Search Index",
 		title: "People mentions as tags",
 		keywords: ["mentions", "people", "tags"],
 	},
 	{
-		id: "advanced-frontmatter",
-		tab: "advanced",
+		id: "general-editor-frontmatter",
+		tab: "general",
 		section: "Editor",
 		title: "Show frontmatter in editor",
 		keywords: ["properties", "yaml", "metadata"],
 	},
 	{
-		id: "advanced-colorful-headings",
-		tab: "advanced",
+		id: "general-editor-colorful-headings",
+		tab: "general",
 		section: "Editor",
 		title: "Colorful headings",
 		keywords: ["editor", "color"],
 	},
 	{
-		id: "advanced-beautiful-tags",
-		tab: "advanced",
-		section: "Editor",
+		id: "appearance-editor-presentation-beautiful-tags",
+		tab: "appearance",
+		section: "Editor presentation",
 		title: "Beautiful Tags",
 		keywords: ["tags", "editor"],
 	},
 	{
-		id: "advanced-editor-width",
-		tab: "advanced",
-		section: "Editor",
+		id: "appearance-editor-presentation-width",
+		tab: "appearance",
+		section: "Editor presentation",
 		title: "Editor width",
 		keywords: ["compact", "comfortable", "wide"],
 	},
 	{
-		id: "advanced-collapsible-headings",
-		tab: "advanced",
+		id: "general-editor-collapsible-headings",
+		tab: "general",
 		section: "Editor",
 		title: "Collapsible headings",
 		keywords: ["fold", "collapse", "expand"],
 	},
 	{
-		id: "advanced-vim-keybindings",
-		tab: "advanced",
+		id: "general-editor-spell-check",
+		tab: "general",
+		section: "Editor",
+		title: "Spell check",
+		keywords: ["spelling", "typo", "dictionary"],
+	},
+	{
+		id: "general-editor-vim-keybindings",
+		tab: "general",
 		section: "Editor",
 		title: "Vim keybindings",
 		keywords: ["vim", "keyboard", "modal"],
 	},
 	{
-		id: "advanced-ai-tools",
-		tab: "advanced",
-		section: "AI",
+		id: "ai-assistant-behavior-tools",
+		tab: "ai",
+		section: "Assistant behavior",
 		title: "AI chat has access to tools",
 		keywords: ["assistant", "tool calls"],
 	},
 	{
-		id: "advanced-folio-mode",
-		tab: "advanced",
-		section: "App",
+		id: "appearance-layout-folio-mode",
+		tab: "appearance",
+		section: "Layout",
 		title: "Folio Mode",
 		keywords: ["layout", "reading"],
 	},
 	{
-		id: "advanced-classic-all-notes",
-		tab: "advanced",
-		section: "App",
+		id: "appearance-layout-classic-all-notes",
+		tab: "appearance",
+		section: "Layout",
 		title: "Classic All Notes grid",
 		keywords: ["all notes", "activity", "timeline", "grid"],
 	},
 	{
-		id: "advanced-folder-counts",
-		tab: "advanced",
-		section: "App",
+		id: "general-file-tree-folder-counts",
+		tab: "general",
+		section: "File tree",
 		title: "Show folder file counts",
 		keywords: ["sidebar", "counts"],
 	},
 	{
-		id: "advanced-show-non-markdown-files",
-		tab: "advanced",
-		section: "App",
+		id: "general-file-tree-non-markdown-files",
+		tab: "general",
+		section: "File tree",
 		title: "Show non-Markdown files",
 		keywords: ["sidebar", "attachments", "pdf", "images", "folio", "file tree"],
 	},
 	{
-		id: "advanced-database-column-color",
-		tab: "advanced",
+		id: "general-file-tree-sort",
+		tab: "general",
+		section: "File tree",
+		title: "File tree sort",
+		keywords: ["sidebar", "folders", "files", "order"],
+	},
+	{
+		id: "appearance-database-column-color",
+		tab: "appearance",
 		section: "Database",
 		title: "Show database column color",
 		keywords: ["collection", "database", "color"],
